+++ game/ato/flightplans/flightplanbuildertypes.py
@@ -5,6 +5,7 @@ from typing import Any, TYPE_CHECKING, Type
 from game.ato import FlightType
 from game.theater.controlpoint import NavalControlPoint
 from .aewc import AewcFlightPlan
+from .afac import AfacFlightPlan
 from .airassault import AirAssaultFlightPlan
 from .airlift import AirliftFlightPlan
 from .antiship import AntiShipFlightPlan
@@ -67,6 +68,7 @@ class FlightPlanBuilderTypes:
             FlightType.PRETENSE_CARGO: PretenseCargoFlightPlan.builder_type(),
             FlightType.ARMED_RECON: ArmedReconFlightPlan.builder_type(),
             FlightType.RECOVERY: RecoveryTankerFlightPlan.builder_type(),
+            FlightType.AFAC: AfacFlightPlan.builder_type(),
         }
         try:
             return builder_dict[flight.flight_type]